backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
python-dotenv
httpx
# psycopg2-binary # Uncomment for PostgreSQL production
